% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_gem}
\alias{clean_gem}
\alias{bind_gem}
\alias{fct_gem}
\title{Clean and tidy raw GEM data}
\usage{
clean_gem(x)

bind_gem(x)

fct_gem(x)
}
\arguments{
\item{x}{a list of GEM data frames}
}
\description{
Clean and tidy raw GEM data
}
