% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_gem_year}
\alias{clean_gem_year}
\alias{clean_gem}
\title{Clean and tidy raw GEM data}
\usage{
clean_gem_year(x)

clean_gem(x)
}
\arguments{
\item{x}{a data frame of raw GEM microdata}

\item{x}{a list of GEM data frames}
}
\description{
Clean and tidy raw GEM data
}
