% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEM-package.R
\docType{package}
\name{GEM-package}
\alias{GEM}
\alias{GEM-package}
\title{GEM: Genomic and Epigenomic Mutation Rate Estimation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://weinstockj.github.io/GEM/}
}

}
\author{
\strong{Maintainer}: Josh Weinstock \email{josh.weinstock@emory.edu}

}
\keyword{internal}
