% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEM_model.R
\name{GEM_Emodel}
\alias{GEM_Emodel}
\title{Emodel Analysis}
\usage{
GEM_Emodel(env_file_name, covariates_file_name, methylation_file_name,
  Emodel_pv, output_file_name, qqplot_file_name)
}
\arguments{
\item{env_file_name}{Phenotype/environmental factor data file}

\item{covariates_file_name}{Covariates data file}

\item{methylation_file_name}{Methylation data file}

\item{Emodel_pv}{Pvalue-cutoff (default is 1.0)}

\item{output_file_name}{output file name}

\item{qqplot_file_name}{qqplot file name}
}
\value{
save results automatically
}
\description{
Emodel is to find the association between methylome and phenotype/environmental factors. The basic function is lm(methylation ~ env + covariates). The output is significance testing on environment.
}
\examples{
DATADIR = system.file('extdata',package='gem')
RESULTDIR = getwd()
env_file_name = paste(DATADIR, "env.txt", sep = "")
covariates_file_name = paste(DATADIR, "cov.txt", sep = "")
covariates_file_name_gxe = paste(DATADIR, "gxe.txt", sep = "")
methylation_file_name = paste(DATADIR, "methylation.txt", sep = "")
snp_file_name = paste(DATADIR, "snp.txt", sep = "")
Emodel_pv = 1
Emodel_result_file_name = paste(RESULTDIR, "Result_Emodel.txt", sep = "")
Emodel_qqplot_file_name = paste(RESULTDIR, "QQplot_Emodel.jpg", sep = "")
GEM_Emodel(env_file_name, covariates_file_name, methylation_file_name, Emodel_pv, Emodel_result_file_name, Emodel_qqplot_file_name)
}

