% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEM_model.R
\name{GEM_Gmodel}
\alias{GEM_Gmodel}
\title{Gmodel Analysis}
\usage{
GEM_Gmodel(snp_file_name, covariates_file_name, methylation_file_name,
  Gmodel_pv, output_file_name)
}
\arguments{
\item{snp_file_name}{Genotype data file}

\item{covariates_file_name}{Covariates data file}

\item{methylation_file_name}{Methylation data file}

\item{Gmodel_pv}{Pvalue-cutoff(default is 1.0e-5 in most cases)}

\item{output_file_name}{output file name}

\item{qqplot_file_name}{qqplot file name}
}
\value{
save results automatically
}
\description{
Gmodel is to find methyQTL that is methylation quantitative trait loci determined by genotyping. The basic function is lm(methylation ~ genotype + covariates). The output is significance testing on genotype.
}
\examples{
DATADIR = system.file('extdata',package='gem')
RESULTDIR = getwd()
env_file_name = paste(DATADIR, "env.txt", sep = "")
covariates_file_name = paste(DATADIR, "cov.txt", sep = "")
covariates_file_name_gxe = paste(DATADIR, "gxe.txt", sep = "")
methylation_file_name = paste(DATADIR, "methylation.txt", sep = "")
snp_file_name = paste(DATADIR, "snp.txt", sep = "")
Gmodel_pv = 1e-04
Gmodel_result_file_name = paste(RESULTDIR, "Result_Gmodel.txt", sep = "")
Gmodel_qqplot_file_name = paste(RESULTDIR, "QQplot_Gmodel.jpg", sep = "")
GEM_Gmodel(snp_file_name, covariates_file_name, methylation_file_name, Gmodel_pv, Gmodel_result_file_name, Gmodel_qqplot_file_name)
}

