% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEM_model.R
\name{GEM_GplusEmodel}
\alias{GEM_GplusEmodel}
\title{GplusEmodel}
\usage{
GEM_GplusEmodel(snp_file_name, covariates_file_name, methylation_file_name,
  GplusEmodel_pv, output_file_name, qqplot_file_name)
}
\arguments{
\item{snp_file_name}{}

\item{covariates_file_name}{}

\item{methylation_file_name}{}

\item{GplusEmodel_pv}{}

\item{output_file_name}{}

\item{qqplot_file_name}{}
}
\value{
save results automatically
}
\description{
GplusEmodel
}

