% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEM_GUI.R
\name{GEM_GUI}
\alias{GEM_GUI}
\title{Graphical User Interface (GUI) for GEM}
\usage{
GEM_GUI()
}
\value{
GEM model analysis results
}
\description{
The user friendly GUI for runing GEM package easily and quickly
}
\details{
The GEM package provides a highly efficient R tool suite for performing epigenome wide association
studies (EWAS). GEM provides three major functions named \code{\link{GEM_Emodel}},
\code{\link{GEM_Gmodel}} and \code{\link{GEM_GxEmodel}} to study the interplay of Gene,
Environment and Methylation (GEM). Within GEM, the pre-existing "Matrix eQTL" package is
utilized and extended to study methylation quantitative trait loci (methQTL) and the
interaction of genotype and environment (GxE) to determine DNA methylation variation,
using matrix based iterative correlation and memory-efficient data analysis.
GEM can facilitate reliable genome-wide methQTL and GxE analysis on a standard laptop
computer within minutes.
}
\examples{
interactive()
#GEM_GUI()  ## remove the hash symbol to run
}
\seealso{
\code{\link{GEM-package}}
}

