% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEM_model.R
\name{GEM_GWASmodel}
\alias{GEM_GWASmodel}
\title{GWASmodel}
\usage{
GEM_GWASmodel(env_file_name, snp_file_name, covariates_file_name, GWASmodel_pv,
  output_file_name, qqplot_file_name)
}
\arguments{
\item{env_file_name}{}

\item{snp_file_name}{}

\item{covariates_file_name}{}

\item{GWASmodel_pv}{}

\item{output_file_name}{}

\item{qqplot_file_name}{}
}
\value{
save results automatically
}
\description{
GWASmodel
}

