% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEM_model.R
\name{GEM_GxEmodel}
\alias{GEM_GxEmodel}
\title{GxEmodel}
\usage{
GEM_GxEmodel(snp_file_name, covariates_file_name, methylation_file_name,
  GxEmodel_pv, output_file_name, topKplot = 10)
}
\arguments{
\item{snp_file_name}{Genotype data file}

\item{covariates_file_name}{Covariates data file}

\item{methylation_file_name}{Methylation data file}

\item{GxEmodel_pv}{Pvalue-cutoff(default is 1.0e-5 in most cases)}

\item{output_file_name}{output file name}

\item{topKplot}{plot the top k number of assocations, if NULL, not plot will be saved}
}
\value{
save results automatically
}
\description{
GxEmodel is to explore how methylome is associated with the interplay between genotype and environmental factor.  The basic function is lm(methylation ~ genotype x environment +  covariates). The output is significance testing on genotype x environment.
}
\examples{
DATADIR = system.file('extdata',package='gem')
RESULTDIR = getwd()
env_file_name = paste(DATADIR, "env.txt", sep = "")
covariates_file_name = paste(DATADIR, "cov.txt", sep = "")
covariates_file_name_gxe = paste(DATADIR, "gxe.txt", sep = "")
methylation_file_name = paste(DATADIR, "methylation.txt", sep = "")
snp_file_name = paste(DATADIR, "snp.txt", sep = "")
GxEmodel_pv = 1
GxEmodel_result_file_name = paste(RESULTDIR, "Result_GxEmodel.txt", sep = "")
GxEmodel_qqplot_file_name = paste(RESULTDIR, "QQplot_GxEmodel.jpg", sep = "")
GEM_GxEmodel(snp_file_name, covariates_file_name_gxe, methylation_file_name, GxEmodel_pv, GxEmodel_result_file_name)
}

