% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Matrix_eQTL_engine.R
\name{Matrix_eQTL_engine2}
\alias{Matrix_eQTL_engine2}
\title{block some cat information, add addInfo}
\usage{
Matrix_eQTL_engine2(snps, gene, cvrt = SlicedData$new(),
  output_file_name = "", pvOutputThreshold = 1e-05,
  useModel = modelLINEAR, errorCovariance = numeric(), verbose = TRUE,
  output_file_name.cis = "", pvOutputThreshold.cis = 0, snpspos = NULL,
  genepos = NULL, cisDist = 1e+06, pvalue.hist = FALSE,
  min.pv.by.genesnp = FALSE, noFDRsaveMemory = FALSE, addInfo = "")
}
\description{
block some cat information, add addInfo
}

