% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gem_data}
\alias{read_gem_data}
\title{Read GEM data from SAV files}
\usage{
read_gem_data(file, ...)
}
\arguments{
\item{file}{path to a SAV file}

\item{arguments}{passed to \code{\link[haven]{read_sav}}
\dontrun{
files <- fs::dir_ls("~/Data/gem", regexp = ".sav$")
gem_data <- read_gem_data(files[1])
}}
}
\description{
Read GEM data from SAV files
}
