% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate simulated data}
\usage{
simulate_data(n = 10000, p = 10, k = 100)
}
\arguments{
\item{n}{Number of candidate mutations}

\item{p}{Number of annotations}

\item{k}{Number of samples}
}
\value{
A list simulated data components
}
\description{
Generate simulated data
}
