% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\alias{test_inSample}
\alias{test_outSample}
\title{Implement gem criteria on a data set}
\usage{
test_inSample(dat, trainingObject)

test_outSample(XFrame, y0, y1, trainingObject)
}
\arguments{
\item{dat}{A dataframe with first column as the treatment index, second column as response, and the
remaining columns as predictors design matrix. Here you should know the number of predictors in the data
set that used to fit the \code{trainingObject} and the \code{dat} should have same number of predictors.}

\item{trainingObject}{A list that return from \code{train}}

\item{XFrame}{Design matrix for the predictors of the out-of-sample data set}

\item{y0}{Response for each observation under the first treatment assignment}

\item{y1}{Response for each observation under the second treatment assignment}
}
\value{
Output of these two functions for the data generator under Gem case is not the same,
and you can find explanation for each of them in the following:
#' \enumerate{
		\item \code{PAB_gem} Population average benefit, see details
		\item \code{es_gem} Effect size of the combined moderator, see details
		\item \code{argumentedData} The input dataframe plus one more column as the combined moderator
		\item \code{combinedModeratorPlot} A ggplot2 object that plots the combined moderator versus response
		with two treatment group marked with different colors
}
}
\description{
Implement a fitted GEM on a testing dataset, calculte the combined moderator, the population
average benefit, the effect size and plot the combined moderator interacting with the treatment.
}
\details{
These are two functions for validating a gem criteria on a data set, for both
in-sample and out-sample types respectively.

The \code{test_inSample} take the data set which
only one reponse is available for each observation and the population average benefit is calculated
as the mean of responses whose pre-assigned treatment coincides with the optimal treatment.

 \code{test_outSample} validate on a data set that two responses for both treatments are available,
and for this function the only output is the population average benefit, which is the mean of the
response under optimal treatment assignmnet

The effect size is calulated in terms of the combined moderator.
}
\examples{
numb_p <-5
bet <- vector("list",2)
bet[[1]] <- c(1,1,1,1,1)
bet[[2]] <- c(1,2,3,4,5)
co <- matrix(0.2,numb_p,numb_p)
      diag(co) <- 1

popu1 <- nongemData_inSample(200,co,0.1,bet,c(0,1))
popu2 <- nongemData_outSample(20000,co,0.2,bet,c(0,1))

mod1 <- train(popu1[[1]],"nu")

test1 <- test_inSample(popu1[[1]],mod1)
test2 <- test_outSample(popu2[[3]],popu2[[1]],popu2[[2]],mod1)
}

